Move osx specific code into osx.go
